+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Mutation `updateClientPreferencesData` to update the client preferences in the order form.
+- Field `clientPreferencesData` to `OrderForm`.
 
 ## [0.26.0] - 2020-03-18