+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.27.0] - 2020-03-19
+
 ### Added
 
 - Mutation `updateClientPreferencesData` to update the client preferences in the order form.